Updated admin index to display all management options. DeleteCinema.php should now correctly display an error message if there are no cinemas in the database. Added AddMovie and DeleteMovie admin pages. They do the things.

+++ DeleteCinema.php
@@ -66,11 +66,12 @@ Delete a Cinema
 		$deleteCinemaQuery = "delete from Cinema where ID=$cinemaID;";								
 		if(mysql_query($deleteCinemaQuery) or die(mysql_error()))
 		{
-			echo "Cinema successfully deleted!";
+			echo "Cinema $cinemaID successfully deleted!";
 		}
 		else
 		{
-			echo "Cinema not found!";
+			echo "Cinema could not be deleted.";
+			echo '<br><a href ="DeleteCinema.php">Go Back</a>';				
 		}
 		
 	}
@@ -78,28 +79,31 @@ Delete a Cinema
 	{
 		$cinemasExistResult = mysql_query("select * from Cinema;") or die(mysql_error());	
 
-		if(mysql_fetch_array($cinemasExistResult))
-		{		
-			echo "<form action='DeleteCinema.php' method='post'>";
-			echo "Delete Cinema: ";
-			echo "<select name='cinemaToDelete'>";
-			
-			$selectCinemaResult = mysql_query("select * from Cinema;") or die(mysql_error());	
-			while($cinemaRow = mysql_fetch_array($selectCinemaResult))
-			{
-					$cinemaID = $cinemaRow['ID'];
-					$cinemaName = $cinemaRow['Name'];
-					$cinemaAddress = $cinemaRow['Address'];
-					
-					echo "<option value='$cinemaID'>($cinemaID) $cinemaName, $cinemaAddress</option>";
-			}
-			
+		if(mysql_num_rows($cinemasExistResult))
+		{
+			if(mysql_fetch_array($cinemasExistResult))
+			{		
+				echo "<form action='DeleteCinema.php' method='post'>";
+				echo "Delete Cinema: ";
+				echo "<select name='cinemaToDelete'>";
+				
+				$selectCinemaResult = mysql_query("select * from Cinema;") or die(mysql_error());	
+				while($cinemaRow = mysql_fetch_array($selectCinemaResult))
+				{
+						$cinemaID = $cinemaRow['ID'];
+						$cinemaName = $cinemaRow['Name'];
+						$cinemaAddress = $cinemaRow['Address'];
+						
+						echo "<option value='$cinemaID'>($cinemaID) $cinemaName, $cinemaAddress</option>";
+				}
+				
 				echo "</select>";
 				echo "<br><br>";
 				echo "<input type='submit' value='Delete Cinema'>";
 				echo "</form>";
-				
+			}
 		}
+		
 		else
 		{
 			echo "No existing cinemas.<br>";
